Rename apps to avoid collisions
