+++ djangae/contrib/gauth/gauth_datastore/permissions.py
