+++ djangae/contrib/gauth/models.py
@@ -9,7 +9,7 @@ database backend
 import warnings
 from exceptions import DeprecationWarning
 
-from djangae.contrib.gauth.datastore.models import (
+from djangae.contrib.gauth.gauth_datastore.models import (
     GaeAbstractDatastoreUser, GaeDatastoreUser)
